more precise input type in normalization

+++ dwi/texture.py
@@ -34,7 +34,7 @@ def normalize(pmap, levels=128):
     # in_range = (0, sp.stats.scoreatpercentile(pmap, 99.8))
     # in_range = (sp.stats.scoreatpercentile(pmap, 0.8),
     #             sp.stats.scoreatpercentile(pmap, 99.2))
-    if pmap.dtype == np.int:
+    if pmap.dtype == np.int32:
         # The rescaler cannot handle int32.
         pmap = np.asarray(pmap, dtype=np.int16)
     if pmap.dtype == np.int16:
